balance works

+++ bank/models.py
@@ -3,6 +3,7 @@ from django.contrib.auth.models import User
 
 
 
+
 # Create your models here.
 
 transaction_type = ['deposit', 'withdrawal']
@@ -16,8 +17,7 @@ class Transaction(models.Model):
         (CREDIT, 'Credit'),
     )
     user = models.ForeignKey(User)
-    dollar_amount = models.FloatField()
+    dollar_amount = models.DecimalField(max_digits=100,decimal_places=2)
     vendor = models.CharField(max_length=50)
     date = models.DateTimeField(auto_now_add=True)
     transaction_type = models.CharField(max_length=10, choices=TRANSACTION_CHOICES, default=DEBIT)
-
